+++ src/models/management-account.js
@@ -51,9 +51,16 @@ ManagementAccount.attributes = {
     modelKey: 'billingState',
     jsonKey: 'billing_state',
   }),
+  emailAddress: Attributes.String({
+    modelKey: 'emailAddress',
+    jsonKey: 'email',
+  }),
   namespaceId: Attributes.String({
     modelKey: 'namespaceId',
     jsonKey: 'namespace_id',
+  }),
+  provider: Attributes.String({
+    modelKey: 'provider',
   }),
   syncState: Attributes.String({
     modelKey: 'syncState',